requirements.txt
albumentations==0.5.2
astunparse==1.6.3
betterexceptions==0.3.3
cachetools==4.2.0
colorama==0.4.4
efficientnetpytorch==0.6.3
flatbuffers==1.12
gast==0.3.3
googleauth==1.24.0
googleauthoauthlib==0.4.2
googlepasta==0.2.0
grpcio==1.32.0
h5py==2.10.0
imgaug==0.4.0
keraspreprocessing==1.1.2
markdown==3.3.3
matplotlib==3.3.3
medpy==0.4.0
munch==2.5.0
nibabel==3.2.1
oauthlib==3.1.0
opencvcontribpython==4.5.1.48
opencvpython==4.5.1.48
opencvpythonheadless==4.5.1.48
opteinsum==3.3.0
pandas==1.2.0
pretrainedmodels==0.7.4
prettyerrors==1.2.19
protobuf==3.14.0
pyasn1==0.4.8
pyasn1modules==0.2.8
pycompare==1.5.1
pydicom==2.1.2
requestsoauthlib==1.3.0
rsa==4.7
scikitlearn==0.24.1
scipy==1.6.0
segmentationmodelspytorch==0.1.3
shapely==1.7.1
simpleitk==2.0.2
tensorboard==2.4.1
tensorboardpluginwit==1.7.0
tensorboardx==2.1
tensorflow==2.4.0
tensorflowestimator==2.4.0
termcolor==1.1.0
timm==0.3.2
tqdm==4.56.0
werkzeug==1.0.1
wrapt==1.12.1
